+++ src/utils/regex.js
@@ -8,7 +8,7 @@ export const contextRegex = /\{\{context:\s?(.*)\}\}|\(\(context:\s?(.*)\)\)/;
 export const templateRegex = /\(\(template:.?(\(\([^\)]{9}\)\))\)\)/;
 export const dateStringRegex = /^[0-9]{2}-[0-9]{2}-[0-9]{4}$/;
 export const numbersRegex = /\d+/g;
-export const roamImageRegex = /!\[[^\]]*\]\((http[^\s)]+)\)/g;
+export const roamImageRegex = /!\[([^\]]*)\]\((http[^\s)]+)\)/g;
 export const sbParamRegex = /^\{.*\}$/;
 export const customTagRegex = {
   "liveai/prompt": /\#?\[?\[?liveai\/prompt\]?\]?/i,