+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **New features**
 
+- Generate or edit images with OpenAI gpt-image-1 model
 - Generate any Mermaid diagrams (and argument map using Mermaid)
 - Text to Speech (using OpenAI gpt-4o-mini-tts)
 - LLMs can generate Roam table or kanban (using 'table' or 'kanban' keyword in your prompt)